+++ tests/include/testing/gl/alg_common.hpp
