+++ tests/source/test_edge_tags.cpp
@@ -1,5 +1,5 @@
-#include "constants.hpp"
-#include "types.hpp"
+#include "testing/gl/constants.hpp"
+#include "testing/gl/types.hpp"
 
 #include <gl/edge_descriptor.hpp>
