+++ tests/source/test_graph.cpp
@@ -1,7 +1,7 @@
-#include "constants.hpp"
-#include "functional.hpp"
-#include "transforms.hpp"
-#include "types.hpp"
+#include "testing/gl/constants.hpp"
+#include "testing/gl/functional.hpp"
+#include "testing/gl/transforms.hpp"
+#include "testing/gl/types.hpp"
 
 #include <gl/graph.hpp>
