+++ tests/source/test_graph_incidence.cpp
@@ -1,5 +1,5 @@
-#include "constants.hpp"
-#include "functional.hpp"
+#include "testing/gl/constants.hpp"
+#include "testing/gl/functional.hpp"
 
 #include <gl/graph.hpp>
