+++ tests/source/test_graph_topology_builders.cpp
@@ -1,4 +1,4 @@
-#include "constants.hpp"
+#include "testing/gl/constants.hpp"
 
 #include <gl/topologies.hpp>
