+++ tests/source/test_non_null_iterator.cpp
@@ -1,5 +1,5 @@
-#include "constants.hpp"
-#include "functional.hpp"
+#include "testing/gl/constants.hpp"
+#include "testing/gl/functional.hpp"
 
 #include <gl/types/iterator_range.hpp>
 #include <gl/types/non_null_iterator.hpp>
