+++ tests/source/test_stream_options_manipulator.cpp
@@ -1,4 +1,4 @@
-#include "constants.hpp"
+#include "testing/gl/constants.hpp"
 
 #include <gl/io/stream_options_manipulator.hpp>
