+++ tests/source/test_vertex_descriptor.cpp
@@ -1,5 +1,5 @@
-#include "constants.hpp"
-#include "types.hpp"
+#include "testing/gl/constants.hpp"
+#include "testing/gl/types.hpp"
 
 #include <gl/vertex_descriptor.hpp>
